refactor: structure of theme component
